Refactored some Controllers to remove unnecessary code/remove abstraction of Author vs Article Controllers
